Move Extension related components to ExtensionPoint dir
