+++ react/components/ExtensionPoint/ExtensionManager.tsx
