+++ react/components/ExtensionPoint/ExtensionPoint.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { cleanup, render } from '@vtex/test-tools/react'
 import 'jest-dom/extend-expect'
-import ExtensionPoint from './ExtensionPoint'
-import { TreePathContextProvider } from '../utils/treePath'
-import { RenderContextProvider } from './RenderContext'
-import { TEST_ID as PreviewTestId } from './Preview'
+import ExtensionPoint from './index'
+import { TreePathContextProvider } from '../../utils/treePath'
+import { RenderContextProvider } from '../RenderContext'
+import { TEST_ID as PreviewTestId } from '../Preview'
 
 afterEach(cleanup)
